IPv6 the CIDR field assigned with wrong type and notation.
For IPv6 networks the netmask is integer and the CIDR field is assigned with `int` to a field with `string` constraint. The netmask in this case represents the number for allow networks.

The IPv6 address `2b03:b0c0:2:d0::102e:d001` should have the CIDR notation `2b03:b0c0:2:d0::102e:d001/64` or `2b03:b0c0:0002:00d0:0000:0000:0000:0000/64`.

I'm not 100% sure what the netmask value can be on the api die, is it always int? or can it come as string? that's why the check for int.